+++ contracts/package.json
@@ -35,10 +35,14 @@
     "typescript": "^5.5",
     "vitest": "^2.1.0"
   },
+  "peerDependencies": {
+    "o1js": "^1.8.0"
+  },
   "engines": {
     "node": ">=18.14.0"
   },
   "dependencies": {
     "o1js-pack": "^0.7.0"
   }
+
 }
