+++ contracts/src/scripts/stress-test.ts
@@ -7,7 +7,12 @@ import {
   UInt64,
   NetworkId,
 } from 'o1js';
-import { NameService, NameRecord, offchainState, Name } from './NameService.js';
+import {
+  NameService,
+  NameRecord,
+  offchainState,
+  Name,
+} from '../NameService.js';
 
 // check command line arg
 let deployAlias = process.argv[2];
@@ -40,16 +45,12 @@ let feepayerKeysBase58: { privateKey: string; publicKey: string } = JSON.parse(
   await fs.readFile(config.feepayerKeyPath, 'utf8')
 );
 
-let zkAppKeysBase58: { privateKey: string; publicKey: string } = JSON.parse(
-  await fs.readFile(config.keyPath, 'utf8')
-);
-
 let cycleNumber = 3;
 let names: string[] = [];
 let nameMap = new Map();
 
 let feepayerKey = PrivateKey.fromBase58(feepayerKeysBase58.privateKey);
-let zkAppKey = PrivateKey.fromBase58(zkAppKeysBase58.privateKey);
+let zkAppKey = PrivateKey.random();
 
 const Network = Mina.Network({
   archive: 'https://api.minascan.io/archive/devnet/v1/graphql',
