+++ contracts/src/utils/get-name.ts
@@ -7,7 +7,7 @@ import {
   UInt64,
   NetworkId,
 } from 'o1js';
-import { NameService, NameRecord, offchainState, Name } from './NameService.js';
+import { NameService, NameRecord, offchainState, Name } from '../NameService.js';
 
 let deployAlias = process.argv[2];
 if (!deployAlias)
